requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-telegram-bot>=21.0
solders>=0.21.0
solana>=0.34.0
base58>=2.1.0
cryptography>=41.0.0
playwright>=1.40.0
httpx>=0.25.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
